Move activities to worker folder
