+++ go/worker/activities/module.go
@@ -3,9 +3,9 @@ package activities
 import (
 	"go.uber.org/fx"
 
-	"github.com/michelangelo-ai/michelangelo/go/components/pipelinerun/utils"
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/cachedoutput"
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/model"
+	"github.com/michelangelo-ai/michelangelo/go/worker/activities/notification"
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/ray"
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/spark"
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/storage"
@@ -19,5 +19,5 @@ var Module = fx.Options(
 	model.Module,
 	cachedoutput.Module,
 	trigger.Module,
-	utils.Module,
+	notification.Module,
 )
