+++ go/worker/activities/notification/BUILD.bazel
@@ -2,15 +2,15 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["notification_utils.go"],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/components/pipelinerun/utils",
+    srcs = [
+        "activities.go",
+        "module.go",
+    ],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/notification",
     visibility = ["//visibility:public"],
     deps = [
-        "//go/base/notification/types:go_default_library",
-        "//proto/api/v2:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//activity:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//workflow:go_default_library",
         "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_zap//:go_default_library",
     ],
@@ -18,12 +18,9 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["notification_utils_test.go"],
+    srcs = ["activities_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//go/base/notification/types:go_default_library",
-        "//proto/api/v2:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
     ],
 )
