+++ go/worker/activities/notification/activities_test.go
@@ -1,0 +1,133 @@
+package notification
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+// TestSendMessageToEmailActivity tests the email activity function.
+//
+// This test verifies that the email activity executes successfully with various
+// request configurations without throwing errors. Since this is a placeholder
+// implementation, it primarily tests the function signature and basic execution flow.
+func TestSendMessageToEmailActivity(t *testing.T) {
+	tests := []struct {
+		name        string
+		request     *SendMessageToEmailActivityRequest
+		description string
+	}{
+		{
+			name: "Valid email request",
+			request: &SendMessageToEmailActivityRequest{
+				To:      []string{"test@uber.com"},
+				Subject: "Test Subject",
+				Text:    "Test message",
+				SendAs:  "michelangelo@uber.com",
+			},
+			description: "Should handle valid email request without error",
+		},
+		{
+			name: "Email request with CC and BCC",
+			request: &SendMessageToEmailActivityRequest{
+				To:      []string{"test@uber.com"},
+				Cc:      []string{"cc@uber.com"},
+				Bcc:     []string{"bcc@uber.com"},
+				Subject: "Test Subject",
+				Text:    "Test message",
+				SendAs:  "sender@uber.com",
+			},
+			description: "Should handle email request with CC and BCC fields",
+		},
+		{
+			name: "Email request with HTML content",
+			request: &SendMessageToEmailActivityRequest{
+				To:      []string{"test@uber.com"},
+				Subject: "Test Subject",
+				HTML:    "<p>HTML test message</p>",
+				SendAs:  "sender@uber.com",
+			},
+			description: "Should handle email request with HTML content",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			ctx := context.Background()
+			err := SendMessageToEmailActivity(ctx, tt.request)
+
+			// Since this is a placeholder implementation, we expect no errors
+			assert.NoError(t, err, tt.description)
+		})
+	}
+}
+
+// TestSendMessageToSlackActivity tests the slack activity function.
+//
+// This test verifies that the slack activity executes successfully with various
+// request configurations without throwing errors.
+func TestSendMessageToSlackActivity(t *testing.T) {
+	tests := []struct {
+		name        string
+		request     *SendMessageToSlackActivityRequest
+		description string
+	}{
+		{
+			name: "Valid slack request",
+			request: &SendMessageToSlackActivityRequest{
+				Channel: "#test-channel",
+				Text:    "Test slack message",
+			},
+			description: "Should handle valid slack request without error",
+		},
+		{
+			name: "Slack request with empty channel",
+			request: &SendMessageToSlackActivityRequest{
+				Channel: "",
+				Text:    "Test message",
+			},
+			description: "Should handle slack request with empty channel",
+		},
+		{
+			name: "Slack request with long message",
+			request: &SendMessageToSlackActivityRequest{
+				Channel: "#notifications",
+				Text:    "This is a very long test message that might exceed typical length limits for testing purposes",
+			},
+			description: "Should handle slack request with long message",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			ctx := context.Background()
+			err := SendMessageToSlackActivity(ctx, tt.request)
+
+			// Since this is a placeholder implementation, we expect no errors
+			assert.NoError(t, err, tt.description)
+		})
+	}
+}
+
+// TestSendMessageToEmailActivityNilRequest tests email activity with nil request.
+func TestSendMessageToEmailActivityNilRequest(t *testing.T) {
+	ctx := context.Background()
+
+	// The current implementation doesn't validate for nil input and will likely panic
+	// This test documents the current behavior - in production we'd want input validation
+	assert.Panics(t, func() {
+		_ = SendMessageToEmailActivity(ctx, nil)
+	}, "Current implementation should panic with nil request - production code should add validation")
+}
+
+// TestSendMessageToSlackActivityNilRequest tests slack activity with nil request.
+func TestSendMessageToSlackActivityNilRequest(t *testing.T) {
+	ctx := context.Background()
+
+	// The current implementation doesn't validate for nil input and will likely panic
+	// This test documents the current behavior - in production we'd want input validation
+	assert.Panics(t, func() {
+		_ = SendMessageToSlackActivity(ctx, nil)
+	}, "Current implementation should panic with nil request - production code should add validation")
+}
