+++ go/worker/workflows/module.go
@@ -1,6 +1,7 @@
 package workflows
 
 import (
+	"github.com/michelangelo-ai/michelangelo/go/worker/workflows/notification"
 	"github.com/michelangelo-ai/michelangelo/go/worker/workflows/ray"
 	"github.com/michelangelo-ai/michelangelo/go/worker/workflows/trigger"
 	"go.uber.org/fx"
@@ -9,4 +10,5 @@ import (
 var Module = fx.Options(
 	ray.Module,
 	trigger.Module,
+	notification.Module,
 )
