+++ go/worker/workflows/notification/BUILD.bazel
@@ -3,16 +3,17 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "notifier.go",
         "module.go",
+        "workflows.go",
     ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/components/pipelinerun/notification",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/workflows/notification",
     visibility = ["//visibility:public"],
     deps = [
         "//go/base/notification/types:go_default_library",
-        "//go/base/workflowclient/interface:go_default_library",
-        "//go/worker/workflows/notification:go_default_library",
+        "//go/worker/activities/notification:go_default_library",
         "//proto/api/v2:go_default_library",
+        "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
+        "@com_github_cadence_workflow_starlark_worker//workflow:go_default_library",
         "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_zap//:go_default_library",
     ],
@@ -20,15 +21,12 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["notifier_test.go"],
+    srcs = ["workflows_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//go/base/workflowclient/interface:go_default_library",
-        "//go/base/workflowclient/interface/interface_mock:go_default_library",
+        "//go/base/notification/types:go_default_library",
         "//proto/api/v2:go_default_library",
-        "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
-        "@org_uber_go_zap//:go_default_library",
     ],
 )
